Refactor widget services and item factories for favorites and empty states.
