+++ app/build.gradle
@@ -408,6 +408,9 @@ dependencies {
     implementation project(':serp-logos-api')
     implementation project(':serp-logos-impl')
 
+    // Widgets
+    implementation "androidx.core:core-remoteviews:1.1.0"
+
     // Deprecated. TODO: Stop using this artifact.
     implementation "androidx.legacy:legacy-support-v4:_"
     debugImplementation Square.leakCanary.android
